+++ lisp/eab-pmacros.el
@@ -178,5 +178,5 @@
  'org-align-list-item
  nil nil
  ""
- "M-U M-j M-h M-SPC M-; - RET M-j M-c M-p M-l M-SPC M-O M-i r M-v 2*SPC
- RET 2*C-w")
+ "M-U M-j M-h M-SPC M-; - RET M-j M-c M-p M-l M-SPC M-O r M-v 2*SPC
+ RET M-SPC M-h M-X M-i 2*C-w")